# api/swagger.yaml
...
swagger: "2.0"
info:
  description: "## TMF API Reference: TMF677 - Usage Consumption\n\n### Release: 18.5\
    \ - December 2018\n\nThis API covers the consumption follow up function providing\
    \ ongoing information about usages related to any subscribed communication products\
    \ (voice, data, TV,…) without having to wait the invoice production. These information\
    \ may concern usages charged on a bucket supervised or not and so the remaining\
    \ credits on the bucket. It allows customers or users to be informed on usages\
    \ done and remaining credits on the buckets that they consume under their purchased\
    \ offers and options.\n\nThe usage consumption API allow to view at a given point\
    \ the balance and the consumption counters of the various buckets (SMS, Voice,\
    \ Data for example) that one or more user(s) consume with each of his devices,\
    \ according to the purchased offers and options.\n\n### Resources \n- **usageConsumptionReport**\
    \ : an usage consumption report retrieves the data related to these balances and\
    \ various consumption counters and calculated at the time of the request by the\
    \ server.\n- **usageConsumptionReportRequest** : used to manage the calculation\
    \ request of an usage consumption report.\n\n### Operations\nThe use of the both\
    \ resources described in this document is conditioned by the capability of the\
...
    \ without necessarily store it.\n- We must also anticipate the use case where\
...
host: "localhost:8080"
basePath: "/tmf-api"
...
schemes:
...
consumes:
...
produces:
...
paths:
  /usageConsumptionReport:
    get:
      tags:
      - "usageConsumptionReport"
      summary: "List or find UsageConsumptionReport objects"
      description: "This operation list or find UsageConsumptionReport entities"
      operationId: "listUsageConsumptionReport"
      parameters:
      - name: "fields"
        in: "query"
        description: "Comma-separated properties to be provided in response"
        required: false
        type: "string"
      - name: "offset"
        in: "query"
        description: "Requested index for start of resources to be provided in response"
        required: false
        type: "integer"
      - name: "limit"
        in: "query"
        description: "Requested number of resources to be provided in response"
        required: false
        type: "integer"
      responses:
        200:
          description: "Success"
          headers:
            X-Total-Count:
              type: "integer"
              description: "Total number of items matching criteria"
            X-Result-Count:
              type: "integer"
              description: "Actual number of items returned in the response body"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/UsageConsumptionReport"
        400:
          description: "Bad Request"
          schema:
            $ref: "#/definitions/Error"
        401:
          description: "Unauthorized"
          schema:
            $ref: "#/definitions/Error"
        403:
          description: "Forbidden"
          schema:
            $ref: "#/definitions/Error"
        404:
          description: "Not Found"
          schema:
            $ref: "#/definitions/Error"
        405:
          description: "Method Not allowed"
          schema:
            $ref: "#/definitions/Error"
        409:
          description: "Conflict"
          schema:
            $ref: "#/definitions/Error"
        500:
          description: "Internal Server Error"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "UsageConsumptionReport"
  /usageConsumptionReport/{id}:
    get:
      tags:
      - "usageConsumptionReport"
      summary: "Retrieves a UsageConsumptionReport by ID"
      description: "This operation retrieves a UsageConsumptionReport entity. Attribute\
        \ selection is enabled for all first level attributes."
      operationId: "retrieveUsageConsumptionReport"
      parameters:
      - name: "id"
        in: "path"
        description: "Identifier of the UsageConsumptionReport"
        required: true
        type: "string"
      - name: "fields"
        in: "query"
        description: "Comma-separated properties to provide in response"
        required: false
        type: "string"
      responses:
        200:
          description: "Success"
          schema:
            $ref: "#/definitions/UsageConsumptionReport"
        400:
          description: "Bad Request"
          schema:
            $ref: "#/definitions/Error"
        401:
          description: "Unauthorized"
          schema:
            $ref: "#/definitions/Error"
        403:
          description: "Forbidden"
          schema:
            $ref: "#/definitions/Error"
        404:
          description: "Not Found"
          schema:
            $ref: "#/definitions/Error"
        405:
          description: "Method Not allowed"
          schema:
            $ref: "#/definitions/Error"
        409:
          description: "Conflict"
          schema:
            $ref: "#/definitions/Error"
        500:
          description: "Internal Server Error"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "UsageConsumptionReport"
    delete:
      tags:
      - "usageConsumptionReport"
      summary: "Deletes a UsageConsumptionReport"
      description: "This operation deletes a UsageConsumptionReport entity."
      operationId: "deleteUsageConsumptionReport"
      parameters:
      - name: "id"
        in: "path"
        description: "Identifier of the UsageConsumptionReport"
        required: true
        type: "string"
      responses:
        204:
          description: "Deleted"
        400:
          description: "Bad Request"
          schema:
            $ref: "#/definitions/Error"
        401:
          description: "Unauthorized"
          schema:
            $ref: "#/definitions/Error"
        403:
          description: "Forbidden"
          schema:
            $ref: "#/definitions/Error"
        404:
          description: "Not Found"
          schema:
            $ref: "#/definitions/Error"
        405:
          description: "Method Not allowed"
          schema:
            $ref: "#/definitions/Error"
        409:
          description: "Conflict"
          schema:
            $ref: "#/definitions/Error"
        500:
          description: "Internal Server Error"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "UsageConsumptionReport"
...
definitions:
...
  UsageConsumptionReport:
    type: "object"
    properties:
      id:
        type: "string"
        description: "The report is calculated at the time of the request by the server.\
          \ Generally, this report is not recorded by the server. If it is, an unique\
          \ identifier of the usage consumption report is given by the server"
      href:
        type: "string"
        description: "Hyperlink to access the usage consumption report. The report\
          \ is calculated at the time of the request by the server. Generally, this\
          \ report is not recorded by the server. If it is, an unique identifier of\
          \ the usage consumption report is given by the server"
      description:
        type: "string"
        description: "Free short text describing the usage consumption report content"
      effectiveDate:
        type: "string"
        format: "date-time"
        description: "Date and time when the usage consumption report was calculated\
          \ and generated"
      name:
        type: "string"
        description: "Usage consumption report name"
      bucket:
        type: "array"
        description: "Bucket(s) included in the offer or option subscribed."
        items:
          $ref: "#/definitions/UsageVolumeProduct"
      relatedParty:
        $ref: "#/definitions/RelatedParty"
      '@baseType':
        type: "string"
        description: "When sub-classing, this defines the super-class"
      '@schemaLocation':
        type: "string"
        format: "uri"
        description: "A URI to a JSON-Schema file that defines additional attributes\
          \ and relationships"
      '@type':
        type: "string"
        description: "When sub-classing, this defines the sub-class entity name"
    description: "An usage consumption report enables to know at a given point the\
      \ balances and the consumption counters related to various buckets (SMS, Voice,\
      \ Data for example). It could be calculated for a device identified by a public\
      \ key (msisdn number for a mobile device for example or PSTN or VOIP number\
      \ for a fix device), for a subscribed offer or option or for an user."
    example:
      bucket:
      - product:
        - '@baseType': "@baseType"
          '@type': "@type"
          name: "name"
          id: "id"
          href: "href"
          '@schemaLocation': "http://example.com/aeiou"
          publicIdentifier: "publicIdentifier"
          user:
          - '@referredType': "@referredType"
            role: "role"
            '@baseType': "@baseType"
            '@type': "@type"
            name: "name"
            id: "id"
            href: "href"
            '@schemaLocation': "@schemaLocation"
          - '@referredType': "@referredType"
            role: "role"
            '@baseType': "@baseType"
            '@type': "@type"
            name: "name"
            id: "id"
            href: "href"
            '@schemaLocation': "@schemaLocation"
          outOfBucketCounter:
          - consumptionPeriod:
              startDateTime: "2000-01-23T04:56:07.000+00:00"
              endDateTime: "2000-01-23T04:56:07.000+00:00"
            product:
              '@referredType': "@referredType"
              '@baseType': "@baseType"
              '@type': "@type"
              name: "name"
              id: "id"
              href: "href"
              '@schemaLocation': "http://example.com/aeiou"
              publicIdentifier: "publicIdentifier"
            '@baseType': "@baseType"
            level: "level"
            valueName: "valueName"
            '@type': "@type"
            '@schemaLocation': "http://example.com/aeiou"
            counterType: "counterType"
            user:
              '@referredType': "@referredType"
              role: "role"
              '@baseType': "@baseType"
              '@type': "@type"
              name: "name"
              id: "id"
              href: "href"
              '@schemaLocation': "@schemaLocation"
            value:
              amount: 0.8008282
              units: "units"
          - consumptionPeriod:
              startDateTime: "2000-01-23T04:56:07.000+00:00"
              endDateTime: "2000-01-23T04:56:07.000+00:00"
            product:
              '@referredType': "@referredType"
              '@baseType': "@baseType"
              '@type': "@type"
              name: "name"
              id: "id"
              href: "href"
              '@schemaLocation': "http://example.com/aeiou"
              publicIdentifier: "publicIdentifier"
            '@baseType': "@baseType"
            level: "level"
            valueName: "valueName"
            '@type': "@type"
            '@schemaLocation': "http://example.com/aeiou"
            counterType: "counterType"
            user:
              '@referredType': "@referredType"
              role: "role"
              '@baseType': "@baseType"
              '@type': "@type"
              name: "name"
              id: "id"
              href: "href"
              '@schemaLocation': "@schemaLocation"
            value:
              amount: 0.8008282
              units: "units"
        - '@baseType': "@baseType"
          '@type': "@type"
          name: "name"
          id: "id"
          href: "href"
          '@schemaLocation': "http://example.com/aeiou"
          publicIdentifier: "publicIdentifier"
          user:
          - '@referredType': "@referredType"
            role: "role"
            '@baseType': "@baseType"
            '@type': "@type"
            name: "name"
            id: "id"
            href: "href"
            '@schemaLocation': "@schemaLocation"
          - '@referredType': "@referredType"
            role: "role"
            '@baseType': "@baseType"
            '@type': "@type"
            name: "name"
            id: "id"
            href: "href"
            '@schemaLocation': "@schemaLocation"
          outOfBucketCounter:
          - consumptionPeriod:
              startDateTime: "2000-01-23T04:56:07.000+00:00"
              endDateTime: "2000-01-23T04:56:07.000+00:00"
            product:
              '@referredType': "@referredType"
              '@baseType': "@baseType"
              '@type': "@type"
              name: "name"
              id: "id"
              href: "href"
              '@schemaLocation': "http://example.com/aeiou"
              publicIdentifier: "publicIdentifier"
            '@baseType': "@baseType"
            level: "level"
            valueName: "valueName"
            '@type': "@type"
            '@schemaLocation': "http://example.com/aeiou"
            counterType: "counterType"
            user:
              '@referredType': "@referredType"
              role: "role"
              '@baseType': "@baseType"
              '@type': "@type"
              name: "name"
              id: "id"
              href: "href"
              '@schemaLocation': "@schemaLocation"
            value:
              amount: 0.8008282
              units: "units"
          - consumptionPeriod:
              startDateTime: "2000-01-23T04:56:07.000+00:00"
              endDateTime: "2000-01-23T04:56:07.000+00:00"
            product:
              '@referredType': "@referredType"
              '@baseType': "@baseType"
              '@type': "@type"
              name: "name"
              id: "id"
              href: "href"
              '@schemaLocation': "http://example.com/aeiou"
              publicIdentifier: "publicIdentifier"
            '@baseType': "@baseType"
            level: "level"
            valueName: "valueName"
            '@type': "@type"
            '@schemaLocation': "http://example.com/aeiou"
            counterType: "counterType"
            user:
              '@referredType': "@referredType"
              role: "role"
              '@baseType': "@baseType"
              '@type': "@type"
              name: "name"
              id: "id"
              href: "href"
              '@schemaLocation': "@schemaLocation"
            value:
              amount: 0.8008282
              units: "units"
        '@baseType': "@baseType"
        bucketCounter:
        - consumptionPeriod:
            startDateTime: "2000-01-23T04:56:07.000+00:00"
            endDateTime: "2000-01-23T04:56:07.000+00:00"
          product:
            '@referredType': "@referredType"
            '@baseType': "@baseType"
            '@type': "@type"
            name: "name"
            id: "id"
            href: "href"
            '@schemaLocation': "http://example.com/aeiou"
            publicIdentifier: "publicIdentifier"
          '@baseType': "@baseType"
          level: "level"
          valueName: "valueName"
          '@type': "@type"
          '@schemaLocation': "http://example.com/aeiou"
          counterType: "counterType"
          user:
            '@referredType': "@referredType"
            role: "role"
            '@baseType': "@baseType"
            '@type': "@type"
            name: "name"
            id: "id"
            href: "href"
            '@schemaLocation': "@schemaLocation"
          value:
            amount: 0.8008282
            units: "units"
        - consumptionPeriod:
            startDateTime: "2000-01-23T04:56:07.000+00:00"
            endDateTime: "2000-01-23T04:56:07.000+00:00"
          product:
            '@referredType': "@referredType"
            '@baseType': "@baseType"
            '@type': "@type"
            name: "name"
            id: "id"
            href: "href"
            '@schemaLocation': "http://example.com/aeiou"
            publicIdentifier: "publicIdentifier"
          '@baseType': "@baseType"
          level: "level"
          valueName: "valueName"
          '@type': "@type"
          '@schemaLocation': "http://example.com/aeiou"
          counterType: "counterType"
          user:
            '@referredType': "@referredType"
            role: "role"
            '@baseType': "@baseType"
            '@type': "@type"
            name: "name"
            id: "id"
            href: "href"
            '@schemaLocation': "@schemaLocation"
          value:
            amount: 0.8008282
            units: "units"
        '@type': "@type"
        bucketBalance:
        - '@baseType': "@baseType"
          validFor:
            startDateTime: "2000-01-23T04:56:07.000+00:00"
            endDateTime: "2000-01-23T04:56:07.000+00:00"
          remainingValue:
            amount: 0.8008282
            units: "units"
          '@type': "@type"
          '@schemaLocation': "http://example.com/aeiou"
          remainingValueName: "remainingValueName"
        - '@baseType': "@baseType"
          validFor:
            startDateTime: "2000-01-23T04:56:07.000+00:00"
            endDateTime: "2000-01-23T04:56:07.000+00:00"
          remainingValue:
            amount: 0.8008282
            units: "units"
          '@type': "@type"
          '@schemaLocation': "http://example.com/aeiou"
          remainingValueName: "remainingValueName"
        name: "name"
        id: "id"
        '@schemaLocation': "http://example.com/aeiou"
        isShared: true
        usageType: "usageType"
      - product:
        - '@baseType': "@baseType"
          '@type': "@type"
          name: "name"
          id: "id"
          href: "href"
          '@schemaLocation': "http://example.com/aeiou"
          publicIdentifier: "publicIdentifier"
          user:
          - '@referredType': "@referredType"
            role: "role"
            '@baseType': "@baseType"
            '@type': "@type"
            name: "name"
            id: "id"
            href: "href"
            '@schemaLocation': "@schemaLocation"
          - '@referredType': "@referredType"
            role: "role"
            '@baseType': "@baseType"
            '@type': "@type"
            name: "name"
            id: "id"
            href: "href"
            '@schemaLocation': "@schemaLocation"
          outOfBucketCounter:
          - consumptionPeriod:
              startDateTime: "2000-01-23T04:56:07.000+00:00"
              endDateTime: "2000-01-23T04:56:07.000+00:00"
            product:
              '@referredType': "@referredType"
              '@baseType': "@baseType"
              '@type': "@type"
              name: "name"
              id: "id"
              href: "href"
              '@schemaLocation': "http://example.com/aeiou"
              publicIdentifier: "publicIdentifier"
            '@baseType': "@baseType"
            level: "level"
            valueName: "valueName"
            '@type': "@type"
            '@schemaLocation': "http://example.com/aeiou"
            counterType: "counterType"
            user:
              '@referredType': "@referredType"
              role: "role"
              '@baseType': "@baseType"
              '@type': "@type"
              name: "name"
              id: "id"
              href: "href"
              '@schemaLocation': "@schemaLocation"
            value:
              amount: 0.8008282
              units: "units"
          - consumptionPeriod:
              startDateTime: "2000-01-23T04:56:07.000+00:00"
              endDateTime: "2000-01-23T04:56:07.000+00:00"
            product:
              '@referredType': "@referredType"
              '@baseType': "@baseType"
              '@type': "@type"
              name: "name"
              id: "id"
              href: "href"
              '@schemaLocation': "http://example.com/aeiou"
              publicIdentifier: "publicIdentifier"
            '@baseType': "@baseType"
            level: "level"
            valueName: "valueName"
            '@type': "@type"
            '@schemaLocation': "http://example.com/aeiou"
            counterType: "counterType"
            user:
              '@referredType': "@referredType"
              role: "role"
              '@baseType': "@baseType"
              '@type': "@type"
              name: "name"
              id: "id"
              href: "href"
              '@schemaLocation': "@schemaLocation"
            value:
              amount: 0.8008282
              units: "units"
        - '@baseType': "@baseType"
          '@type': "@type"
          name: "name"
          id: "id"
          href: "href"
          '@schemaLocation': "http://example.com/aeiou"
          publicIdentifier: "publicIdentifier"
          user:
          - '@referredType': "@referredType"
            role: "role"
            '@baseType': "@baseType"
            '@type': "@type"
            name: "name"
            id: "id"
            href: "href"
            '@schemaLocation': "@schemaLocation"
          - '@referredType': "@referredType"
            role: "role"
            '@baseType': "@baseType"
            '@type': "@type"
            name: "name"
            id: "id"
            href: "href"
            '@schemaLocation': "@schemaLocation"
          outOfBucketCounter:
          - consumptionPeriod:
              startDateTime: "2000-01-23T04:56:07.000+00:00"
              endDateTime: "2000-01-23T04:56:07.000+00:00"
            product:
              '@referredType': "@referredType"
              '@baseType': "@baseType"
              '@type': "@type"
              name: "name"
              id: "id"
              href: "href"
              '@schemaLocation': "http://example.com/aeiou"
              publicIdentifier: "publicIdentifier"
            '@baseType': "@baseType"
            level: "level"
            valueName: "valueName"
            '@type': "@type"
            '@schemaLocation': "http://example.com/aeiou"
            counterType: "counterType"
            user:
              '@referredType': "@referredType"
              role: "role"
              '@baseType': "@baseType"
              '@type': "@type"
              name: "name"
              id: "id"
              href: "href"
              '@schemaLocation': "@schemaLocation"
            value:
              amount: 0.8008282
              units: "units"
          - consumptionPeriod:
              startDateTime: "2000-01-23T04:56:07.000+00:00"
              endDateTime: "2000-01-23T04:56:07.000+00:00"
            product:
              '@referredType': "@referredType"
              '@baseType': "@baseType"
              '@type': "@type"
              name: "name"
              id: "id"
              href: "href"
              '@schemaLocation': "http://example.com/aeiou"
              publicIdentifier: "publicIdentifier"
            '@baseType': "@baseType"
            level: "level"
            valueName: "valueName"
            '@type': "@type"
            '@schemaLocation': "http://example.com/aeiou"
            counterType: "counterType"
            user:
              '@referredType': "@referredType"
              role: "role"
              '@baseType': "@baseType"
              '@type': "@type"
              name: "name"
              id: "id"
              href: "href"
              '@schemaLocation': "@schemaLocation"
            value:
              amount: 0.8008282
              units: "units"
        '@baseType': "@baseType"
        bucketCounter:
        - consumptionPeriod:
            startDateTime: "2000-01-23T04:56:07.000+00:00"
            endDateTime: "2000-01-23T04:56:07.000+00:00"
          product:
            '@referredType': "@referredType"
            '@baseType': "@baseType"
            '@type': "@type"
            name: "name"
            id: "id"
            href: "href"
            '@schemaLocation': "http://example.com/aeiou"
            publicIdentifier: "publicIdentifier"
          '@baseType': "@baseType"
          level: "level"
          valueName: "valueName"
          '@type': "@type"
          '@schemaLocation': "http://example.com/aeiou"
          counterType: "counterType"
          user:
            '@referredType': "@referredType"
            role: "role"
            '@baseType': "@baseType"
            '@type': "@type"
            name: "name"
            id: "id"
            href: "href"
            '@schemaLocation': "@schemaLocation"
          value:
            amount: 0.8008282
            units: "units"
        - consumptionPeriod:
            startDateTime: "2000-01-23T04:56:07.000+00:00"
            endDateTime: "2000-01-23T04:56:07.000+00:00"
          product:
            '@referredType': "@referredType"
            '@baseType': "@baseType"
            '@type': "@type"
            name: "name"
            id: "id"
            href: "href"
            '@schemaLocation': "http://example.com/aeiou"
            publicIdentifier: "publicIdentifier"
          '@baseType': "@baseType"
          level: "level"
          valueName: "valueName"
          '@type': "@type"
          '@schemaLocation': "http://example.com/aeiou"
          counterType: "counterType"
          user:
            '@referredType': "@referredType"
            role: "role"
            '@baseType': "@baseType"
            '@type': "@type"
            name: "name"
            id: "id"
            href: "href"
            '@schemaLocation': "@schemaLocation"
          value:
            amount: 0.8008282
            units: "units"
        '@type': "@type"
        bucketBalance:
        - '@baseType': "@baseType"
          validFor:
            startDateTime: "2000-01-23T04:56:07.000+00:00"
            endDateTime: "2000-01-23T04:56:07.000+00:00"
          remainingValue:
            amount: 0.8008282
            units: "units"
          '@type': "@type"
          '@schemaLocation': "http://example.com/aeiou"
          remainingValueName: "remainingValueName"
        - '@baseType': "@baseType"
          validFor:
            startDateTime: "2000-01-23T04:56:07.000+00:00"
            endDateTime: "2000-01-23T04:56:07.000+00:00"
          remainingValue:
            amount: 0.8008282
            units: "units"
          '@type': "@type"
          '@schemaLocation': "http://example.com/aeiou"
          remainingValueName: "remainingValueName"
        name: "name"
        id: "id"
        '@schemaLocation': "http://example.com/aeiou"
        isShared: true
        usageType: "usageType"
      '@baseType': "@baseType"
      '@type': "@type"
      name: "name"
      description: "description"
      id: "id"
      href: "href"
      relatedParty:
        '@referredType': "@referredType"
        role: "role"
        '@baseType': "@baseType"
        '@type': "@type"
        name: "name"
        id: "id"
        href: "href"
        '@schemaLocation': "@schemaLocation"
      '@schemaLocation': "http://example.com/aeiou"
      effectiveDate: "2000-01-23T04:56:07.000+00:00"
...